requirements/base.txt
Flask==0.9
Flask-SQLAlchemy==0.16
psycopg2>=2.4.5
